refactor(runtime): split workspace helpers into SRP submodules (#0000)
